+++ apps/bare-expo/modules/benchmarking/ios/BenchmarkingExpoModule.swift
@@ -1,0 +1,17 @@
+import ExpoModulesCore
+
+public final class BenchmarkingExpoModule: Module {
+  public func definition() -> ModuleDefinition {
+    Name("BenchmarkingExpoModule")
+
+    Function("nothing") {}
+
+    Function("addNumbers") { (a: Double, b: Double) in
+      return a + b
+    }
+
+    Function("addStrings") { (a: String, b: String) in
+      return a + b
+    }
+  }
+}