+++ apps/router-e2e/__e2e__/dom-components/components/02-actions.tsx
@@ -14,6 +14,7 @@ export default function Page({
   showAlert: (time: number) => void;
   throwError: () => never;
   getNativeSettings: () => Promise<string>;
+  dom?: import('expo/dom').DOMProps;
 }) {
   const [nativeSetting, setNativeSetting] = useState<string | null>(null);