+++ packages/@expo/cli/src/prebuild/resolveOptions.ts
@@ -8,6 +8,13 @@ import * as Log from '../log';
 import { CommandError } from '../utils/errors';
 import { validateUrl } from '../utils/url';
 
+const debug = require('debug')('expo:prebuild:resolveOptions') as typeof console.log;
+
+export interface ResolvedTemplateOption {
+  type: 'file' | 'npm' | 'repository';
+  uri: string;
+}
+
 export function resolvePackageManagerOptions(args: any) {
   const managers: Record<string, boolean> = {
     npm: args['--npm'],
@@ -31,18 +38,49 @@ export function resolvePackageManagerOptions(args: any) {
 }
 
 /** Resolves a template option as a URL or file path pointing to a tar file. */
-export function resolveTemplateOption(template: string) {
-  if (validateUrl(template)) {
-    return template;
-  }
-  const templatePath = path.resolve(template);
-  assert(fs.existsSync(templatePath), 'template file does not exist: ' + templatePath);
-  assert(
-    fs.statSync(templatePath).isFile(),
-    'template must be a tar file created by running `npm pack` in a project: ' + templatePath
-  );
+export function resolveTemplateOption(template: string): ResolvedTemplateOption {
+  assert(template, 'template is required');
 
-  return templatePath;
+  if (template.startsWith('https://') || template.startsWith('http://')) {
+    if (!validateUrl(template)) {
+      throw new CommandError('BAD_ARGS', 'Invalid URL provided as a template');
+    }
+    debug('Resolved template to repository path:', template);
+    return { type: 'repository', uri: template };
+  }
+
+  if (
+    // Supports `file:./path/to/template.tgz`
+    template.startsWith('file:') ||
+    // Supports `../path/to/template.tgz`
+    template.startsWith('.') ||
+    // Supports `\\path\\to\\template.tgz`
+    template.startsWith(path.sep)
+  ) {
+    let resolvedUri = template;
+    if (resolvedUri.startsWith('file:')) {
+      resolvedUri = resolvedUri.substring(5);
+    }
+    const templatePath = path.resolve(resolvedUri);
+    assert(fs.existsSync(templatePath), 'template file does not exist: ' + templatePath);
+    assert(
+      fs.statSync(templatePath).isFile(),
+      'template must be a tar file created by running `npm pack` in a project: ' + templatePath
+    );
+
+    debug(`Resolved template to file path:`, templatePath);
+    return { type: 'file', uri: templatePath };
+  }
+
+  if (fs.existsSync(template)) {
+    // Backward compatible with the old local template argument, e.g. `--template dir/template.tgz`
+    const templatePath = path.resolve(template);
+    debug(`Resolved template to file path:`, templatePath);
+    return { type: 'file', uri: templatePath };
+  }
+
+  debug(`Resolved template to NPM package:`, template);
+  return { type: 'npm', uri: template };
 }
 
 /** Resolves dependencies to skip from a string joined by `,`. Example: `react-native,expo,lodash` */