+++ packages/babel-preset-expo/src/__tests__/use-dom-directive-plugin.test.ts
@@ -41,15 +41,10 @@ afterAll(() => {
   process.env = { ...originalEnv };
 });
 
-function transformClient(
-  sourceCode: string,
-  platform: string = 'ios',
-  isDev: boolean = false,
-  isDOM: boolean = false
-) {
+function transformClient(sourceCode: string, platform: string = 'ios', isDev: boolean = false) {
   const options = {
     ...DEF_OPTIONS,
-    caller: getCaller({ ...ENABLED_CALLER, isReactServer: false, platform, isDev, isDOM }),
+    caller: getCaller({ ...ENABLED_CALLER, isReactServer: false, platform, isDev }),
   };
 
   const results = babel.transform(sourceCode, options);
@@ -93,41 +88,6 @@ export default function App() {
       });"
     `);
   });
-});
-
-it(`registers a DOM component on web for dom bundles`, () => {
-  const res = transformClient(sourceCode, 'web', true, true);
-  expect(res.code).toMatch('react');
-  expect(res.code).toMatch('expo/dom/internal');
-  expect(res.code).toMatch('registerDOMComponent');
-  expect(res.code).toMatchInlineSnapshot(`
-    ""use dom";
-
-    var _jsxFileName = "/unknown";
-    import { jsxDEV as _jsxDEV } from "react/jsx-dev-runtime";
-    export default function App() {
-      return /*#__PURE__*/_jsxDEV("div", {}, void 0, false, {
-        fileName: _jsxFileName,
-        lineNumber: 5,
-        columnNumber: 10
-      }, this);
-    }
-    require('expo/dom/internal').registerDOMComponent(exports.default);"
-  `);
-});
-
-it(`skips registering a DOM component on web for dom bundles without directive`, () => {
-  const res = transformClient(
-    `export default function App() {
-  return <div />
-}`,
-    'web',
-    true,
-    true
-  );
-  expect(res.code).toMatch('react');
-  expect(res.code).not.toMatch('expo/dom/internal');
-  expect(res.code).not.toMatch('registerDOMComponent');
 });
 
 it(`does nothing on web`, () => {