+++ packages/expo-file-system/ios/Next/FileSystemNextExceptions.swift
@@ -12,3 +12,15 @@ internal class MoveDirectoryToFileException: Exception {
     "Unable to move a directory to a file"
   }
 }
+
+internal class UnableToDownloadException: GenericException<String> {
+  override var reason: String {
+    "Unable to download a file: \(param)"
+  }
+}
+
+internal class UnableToGetFileSizeException: GenericException<String> {
+  override var reason: String {
+    "Unable to get file size: \(param)"
+  }
+}