+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/types/AnyType.kt
@@ -1,7 +1,28 @@
 package expo.modules.kotlin.types
 
+import android.net.Uri
+import com.facebook.react.bridge.ReadableArray
+import com.facebook.react.bridge.ReadableMap
+import expo.modules.core.arguments.ReadableArguments
 import expo.modules.kotlin.AppContext
 import expo.modules.kotlin.jni.ExpectedType
+import expo.modules.kotlin.jni.JavaScriptObject
+import expo.modules.kotlin.jni.JavaScriptValue
+import expo.modules.kotlin.typedarray.BigInt64Array
+import expo.modules.kotlin.typedarray.BigUint64Array
+import expo.modules.kotlin.typedarray.Float32Array
+import expo.modules.kotlin.typedarray.Float64Array
+import expo.modules.kotlin.typedarray.Int16Array
+import expo.modules.kotlin.typedarray.Int32Array
+import expo.modules.kotlin.typedarray.Int8Array
+import expo.modules.kotlin.typedarray.TypedArray
+import expo.modules.kotlin.typedarray.Uint16Array
+import expo.modules.kotlin.typedarray.Uint32Array
+import expo.modules.kotlin.typedarray.Uint8Array
+import expo.modules.kotlin.typedarray.Uint8ClampedArray
+import java.io.File
+import java.net.URI
+import java.net.URL
 import kotlin.reflect.KClass
 import kotlin.reflect.KType
 import kotlin.reflect.KTypeProjection
@@ -44,6 +65,86 @@ class LazyKType(
   }
 }
 
+class EmptyKType(
+  override val classifier: KClass<*>,
+  override val isMarkedNullable: Boolean = false
+) : KType {
+  override val annotations: List<Annotation>
+    get() = emptyList()
+  override val arguments: List<KTypeProjection>
+    get() = emptyList()
+
+  override fun equals(other: Any?): Boolean {
+    if (this === other) return true
+    if (other !is EmptyKType) return false
+
+    return classifier == other.classifier && isMarkedNullable == other.isMarkedNullable
+  }
+
+  override fun hashCode(): Int {
+    return 31 * classifier.hashCode() + isMarkedNullable.hashCode()
+  }
+}
+
+object AnyTypeProvider {
+  @PublishedApi
+  internal val typesMap = buildMap<Pair<KClass<*>, Boolean>, AnyType> {
+    listOf(
+      Int::class,
+      Float::class,
+      Double::class,
+      Long::class,
+      Boolean::class,
+      String::class,
+
+      ByteArray::class,
+      LongArray::class,
+      IntArray::class,
+      BooleanArray::class,
+      FloatArray::class,
+      DoubleArray::class,
+
+      JavaScriptValue::class,
+      JavaScriptObject::class,
+
+      TypedArray::class,
+      Int8Array::class,
+      Int16Array::class,
+      Int32Array::class,
+      Uint8Array::class,
+      Uint8ClampedArray::class,
+      Uint16Array::class,
+      Uint32Array::class,
+      Float32Array::class,
+      Float64Array::class,
+      BigInt64Array::class,
+      BigUint64Array::class,
+
+      ReadableArray::class,
+      ReadableMap::class,
+
+      URL::class,
+      Uri::class,
+      URI::class,
+
+      File::class,
+
+      Any::class,
+      Unit::class,
+
+      ReadableArguments::class
+    ).forEach { klass ->
+      put((klass to false), AnyType(EmptyKType(klass, false)))
+      put((klass to true), AnyType(EmptyKType(klass, true)))
+    }
+  }
+
+  inline fun <reified T> cachedAnyType(): AnyType? {
+    val key = Pair(T::class, null is T)
+    return typesMap[key]
+  }
+}
+
 inline fun <reified T> (() -> KType).toAnyType() = AnyType(
   LazyKType(
     classifier = T::class,
@@ -53,7 +154,7 @@ inline fun <reified T> (() -> KType).toAnyType() = AnyType(
 )
 
 inline fun <reified T> toAnyType(): AnyType {
-  return { typeOf<T>() }.toAnyType<T>()
+  return AnyTypeProvider.cachedAnyType<T>() ?: { typeOf<T>() }.toAnyType<T>()
 }
 
 @Suppress("UNUSED_PARAMETER")
@@ -192,7 +293,8 @@ class AnyType(
     TypeConverterProviderImpl.obtainTypeConverter(kType)
   }
 
-  fun convert(value: Any?, appContext: AppContext? = null): Any? = converter.convert(value, appContext)
+  fun convert(value: Any?, appContext: AppContext? = null): Any? =
+    converter.convert(value, appContext)
 
   fun getCppRequiredTypes(): ExpectedType = converter.getCppRequiredTypes()