+++ packages/expo-video/android/src/main/java/expo/modules/video/VideoModule.kt
@@ -4,9 +4,11 @@ package expo.modules.video
 
 import android.app.Activity
 import android.net.Uri
+import androidx.media3.common.C
 import androidx.media3.common.PlaybackParameters
 import androidx.media3.common.Player.REPEAT_MODE_OFF
 import androidx.media3.common.Player.REPEAT_MODE_ONE
+import androidx.media3.common.Timeline
 import com.facebook.react.uimanager.PixelUtil
 import com.facebook.react.uimanager.Spacing
 import com.facebook.react.uimanager.ViewProps
@@ -50,7 +52,6 @@ class VideoModule : Module() {
 
       Prop("player") { view: VideoView, player: VideoPlayer ->
         view.videoPlayer = player
-        player.prepare()
       }
 
       Prop("nativeControls") { view: VideoView, useNativeControls: Boolean ->
@@ -151,7 +152,11 @@ class VideoModule : Module() {
 
     Class(VideoPlayer::class) {
       Constructor { source: VideoSource? ->
-        VideoPlayer(activity.applicationContext, appContext, source)
+        val player = VideoPlayer(activity.applicationContext, appContext, source)
+        appContext.mainQueue.launch {
+          player.prepare()
+        }
+        return@Constructor player
       }
 
       Property("playing")
@@ -192,6 +197,34 @@ class VideoModule : Module() {
         .set { ref: VideoPlayer, currentTime: Double ->
           appContext.mainQueue.launch {
             ref.player.seekTo((currentTime * 1000).toLong())
+          }
+        }
+
+      Property("currentLiveTimestamp")
+        .get { ref: VideoPlayer ->
+          // TODO: same as `currentTime`
+          runBlocking(appContext.mainQueue.coroutineContext) {
+            val window = Timeline.Window()
+            if (!ref.player.currentTimeline.isEmpty) {
+              ref.player.currentTimeline.getWindow(ref.player.currentMediaItemIndex, window)
+            }
+            if (window.windowStartTimeMs == C.TIME_UNSET) {
+              null
+            } else {
+              window.windowStartTimeMs + ref.player.currentPosition
+            }
+          }
+        }
+
+      Property("currentOffsetFromLive")
+        .get { ref: VideoPlayer ->
+          // TODO: same as `currentTime`
+          runBlocking(appContext.mainQueue.coroutineContext) {
+            if (ref.player.currentLiveOffset == C.TIME_UNSET) {
+              null
+            } else {
+              ref.player.currentLiveOffset / 1000f
+            }
           }
         }
 
@@ -286,9 +319,7 @@ class VideoModule : Module() {
 
         appContext.mainQueue.launch {
           ref.uncommittedSource = videoSource
-          if (VideoManager.isVideoPlayerAttachedToView(ref)) {
-            ref.prepare()
-          }
+          ref.prepare()
         }
       }