+++ packages/expo/src/dom/dom.ts
@@ -1,6 +1,6 @@
 // Native file
 
-export type DOMProps = Omit<import('react-native-webview').WebViewProps, 'source'>;
+export type { DOMProps } from './dom.types';
 
 // TODO: Maybe this could be a bundler global instead.
 /** @returns `true` when the current JS running in a DOM Component environment. */